products/d-manager/requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
requests>=2.31.0
